requirements.txt
flask
flask-socketio
neurosity
python-dotenv
pandas
numpy
mediapipe
opencv-python
